// Angular/src/app/Models/employee.ts
export class Employee {
  id: number = 0;
  name: string = '';
  salary: number = 0;
  attendingTime: string =''
  leavingTime: string =''
  address: string = '';
  gender: string = '';
  dateOfBirth: Date ;
  phoneNumber: string = '';
  nationality: string = '';
  nationalId: string = '';
  dateOfContract: Date ;
  extraHour: number = 0;
  penaltyHour: number = 0;

}